(fix): scope the miner_res variable
It seems that the miner_res variable is not scoped via let or const
keyword. It is simply defined dynamically. This way the miner_res
variable is defined as a global variable. As a result, it could happend
when 2 devices are reporting at the same time, the miner_res variable is
overridden by the next request and let the first request fail because it
has simply the wrong miner_res variable.

By scoping this var via `let` it is scoped to the function and therefor
it cannot be override by the other requests.